backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
feedparser==6.0.11
newspaper3k==0.2.8
httpx==0.27.2
apscheduler==3.10.4
anthropic==0.34.2
python-dotenv==1.0.1
lxml[html_clean]==5.3.0
lxml-html-clean
Pillow==10.4.0
nltk==3.9.1
